Support floatingpoint CO2

+++ Sequor/Models/Wallet.swift
@@ -3,5 +3,5 @@ struct Wallet: Codable {
   /// UserId of the owner of the wallet. (hex)
   let userId: String
   /// Total co2 in KG saved by the user.
-  let totalCO2: Int
+  let totalCO2: Double
 }
